src/backend/requirements.txt
Django
djangorestframework
django-cors-headers
numpy
WhiteNoise
requests
python-dotenv